+++ MP4_FTYP.cpp
@@ -30,9 +30,9 @@ CMP4_FTYP::CMP4_FTYP(void)
 CMP4_FTYP::~CMP4_FTYP(void)
 {
 }
-void CMP4_FTYP::load(FILE *Stream,  u32 offset, u32 size)
+void CMP4_FTYP::load(FILE *Stream,  u64 offset, u64 size)
 {
-	_fseeki64(Stream, offset, SEEK_SET);
+	_fseeki64(Stream, (__int64)offset, SEEK_SET);
 	_blob.FileRead(size, Stream);
 	if (_blob.GetLength() > 8)
 	{
